deal_flow_ingest/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deal-flow-ingest"
version = "0.1.0"
description = "Ingestion-first MVP for Deal Flow"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.0",
  "requests>=2.31",
  "SQLAlchemy>=2.0",
  "pydantic>=2.0",
  "PyYAML>=6.0",
  "alembic>=1.13",
  "openpyxl>=3.1",
  "pyshp>=3.0",
  "streamlit>=1.42",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
dealflow = "deal_flow_ingest.dealflow:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["deal_flow_ingest*"]
